requirements.txt
PyYAML
requests
plantuml
six
jsonschema
pytest
wavedrom
ruamel.yaml
svgwrite
